[10.0][PORT] pos_product_category_discount (#493)
* [FIX] discount rounding after refresh page

* [REF] refactoring code, [ADD] new technical module

* [FIX] client discount after change client
