+++ pos_discount_base/doc/changelog.rst
@@ -1,0 +1,4 @@
+`1.0.0`
+-------
+
+- Init version
