+++ pos_discount_base/doc/index.rst
@@ -1,0 +1,10 @@
+===================
+ POS Discount Base
+===================
+
+Technical module in POS.
+
+Installation
+============
+
+* `Install <https://odoo-development.readthedocs.io/en/latest/odoo/usage/install-module.html>`__ this module in a usual way
