+++ pos_discount_base/static/description/index.html
@@ -1,0 +1,23 @@
+<section class="oe_container">
+  <div class="oe_row oe_spaced">
+	<div class="oe_span12">
+	  <h2 class="oe_slogan">POS discount base</h2>
+	  <h3 class="oe_slogan">Technical module to make POS Discount more extendable</h3>
+    </div>
+  </div>
+</section>
+
+<section class="oe_container">
+  <div class="row mt32 o_animate o_animate_in_children o_animate_offset_min o_visible o_animating" style="animation-name: none; visibility: visible; animation-play-state: running;">
+    <div class="oe_row oe_spaced" style="transition-delay: 0ms;">
+      <div class="oe_span12">
+        <h2>Need our service?</h2>
+        <p class="oe_mt32">Contact us by <a href="mailto:it@it-projects.info">email</a> or fill out <a href="https://www.it-projects.info/page/website.contactus " target="_blank">request form</a></p>
+        <ul>
+          <li><a href="mailto:it@it-projects.info">it@it-projects.info <i class="fa fa-envelope-o"></i></a></li>
+          <li><a href="https://www.it-projects.info/page/website.contactus " target="_blank">https://www.it-projects.info/page/website.contactus <i class="fa fa-list-alt"></i></a></li>
+        </ul>
+      </div>
+    </div>
+  </div>
+</section>
