+++ pos_discount_base/views/template.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <template id="assets_backend" name="POS restaurant base" inherit_id="point_of_sale.assets">
+            <xpath expr="." position="inside">
+                <script type="text/javascript" src="/pos_discount_base/static/lib/js/discount.js"></script>
+            </xpath>
+        </template>
+    </data>
+</openerp>
