+++ pos_product_category_discount/__manifest__.py
@@ -5,30 +5,31 @@
     "category": "Point of Sale",
     "live_test_url": 'http://apps.it-projects.info/shop/product/pos-product-category-discount?version=10.0',
     "images": ['images/dp3.png'],
-    "version": "1.1.0",
+    "version": "1.1.1",
     "application": False,
 
     "author": "IT-Projects LLC, Dinar Gabbasov",
     "support": "apps@it-projects.info",
-    "website": "https://twitter.com/gabbasov_dinar",
+    "website": "https://it-projects.info/team/GabbasovDinar",
     "license": "LGPL-3",
     "price": 39.00,
     "currency": "EUR",
 
     "depends": [
-        'point_of_sale',
-        'pos_discount',
+        "pos_discount_base",
     ],
     "external_dependencies": {"python": [], "bin": []},
     "data": [
-        'views/template.xml',
-        'views/view.xml',
-        'security/ir.model.access.csv',
+        "views/template.xml",
+        "views/view.xml",
+        "security/ir.model.access.csv",
     ],
-    'qweb': [
-        'static/src/xml/DiscountProgram.xml',
+    "qweb": [
+        "static/src/xml/DiscountProgram.xml",
     ],
-    "demo": [],
+    "demo": [
+        "data/pos_product_category_discount_demo.xml",
+    ],
 
     "post_load": None,
     "pre_init_hook": None,
