+++ pos_product_category_discount/doc/changelog.rst
@@ -3,6 +3,11 @@
 Updates
 =======
 
+`1.1.1`
+------
+
+- **FIX:** ability to work in offline mode
+
 `1.1.0`
 ______
