+++ pos_product_category_discount/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_default
